requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
djangorestframework==3.14.0
djangorestframework-camel-case==1.4.2
docopt==0.6.2
idna==3.4
psycopg==3.1.9
psycopg-binary==3.1.9
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
yarg==0.1.9
